requirements.txt
screepsapi==0.5.1
pymongo==3.12.0
PyYAML==5.4.1